+++ src/lib/viz/dataview/mode/DataViewLocal.ts
@@ -1,16 +1,15 @@
-import { Layer, GeoJSONSource } from '@/viz';
+import { Layer } from '@/viz';
 import { AggregationType, aggregate } from '@/data/operations/aggregation/aggregation';
 import { groupValuesByColumn } from '@/data/operations/grouping';
 import { castToNumberOrUndefined } from '@/core/utils/number';
+import { ColumnFilters, SpatialFilters } from '@/viz/filters/types';
 import { DataViewMode } from './DataViewMode';
 
 export class DataViewLocal extends DataViewMode {
-  private useViewport = true;
+  private useViewport = false;
 
-  constructor(dataSource: Layer, column: string, useViewport = true) {
+  constructor(dataSource: Layer, column: string) {
     super(dataSource, column);
-
-    this.useViewport = useViewport;
     this.bindEvents();
   }
 
@@ -21,18 +20,15 @@ export class DataViewLocal extends DataViewMode {
 
     const { excludedFilters = [] } = options;
 
-    if (this.useViewport) {
-      await (this.dataSource as Layer).addSourceField(this.column);
-      return (this.dataSource as Layer).getViewportFeatures(excludedFilters);
-    }
-
-    // is GeoJSON Layer
     if (this.dataSource instanceof Layer) {
-      return (this.dataSource.source as GeoJSONSource).getFeatures(excludedFilters);
+      await this.dataSource.addSourceField(this.column);
+    } else {
+      this.dataSource.addField(this.column);
     }
 
-    // is GeoJSON Source
-    return (this.dataSource as GeoJSONSource).getFeatures(excludedFilters);
+    return this.useViewport
+      ? (this.dataSource as Layer).getViewportFeatures(excludedFilters)
+      : this.dataSource.getFeatures(excludedFilters);
   }
 
   public async groupBy(
@@ -51,6 +47,16 @@ export class DataViewLocal extends DataViewMode {
       .sort(categoryOrder());
 
     return { nullCount, categories };
+  }
+
+  public setFilters(filters: ColumnFilters) {
+    if (this.dataSource instanceof Layer) {
+      this.dataSource.setFilters(filters);
+    }
+  }
+
+  public setSpatialFilter(spatialFilter: SpatialFilters) {
+    this.useViewport = spatialFilter === 'viewport';
   }
 
   private bindEvents() {